+++ pkg/initdir/initdir.go
@@ -1,4 +1,4 @@
-package init
+package initdir
 
 import (
 	"fmt"
