+++ pkg/initdir/initdir_test.go
@@ -1,4 +1,4 @@
-package init
+package initdir
 
 import (
 	"os"
